# base64: some optimizations