--- a/vlib/encoding/base64/base64_memory_test.v
+++ b/vlib/encoding/base64/base64_memory_test.v
@@ -1,26 +1,32 @@
 import encoding.base64
 
 fn test_long_encoding(){
-	repeats := 256
+	repeats := 128
 	input_size := 131072
 	
 	s_original := 'a'.repeat(input_size)
-	mut s_encoded := base64.encode(s_original)
+	s_encoded := base64.encode(s_original)
+	s_decoded := base64.decode(s_encoded)
+	
 	assert s_encoded.len > s_original.len
-	mut s_decoded := base64.decode(s_encoded)
 	assert s_original == s_decoded
 
 	mut s := 0
+	
+	ebuffer := malloc( s_encoded.len )
 	for i := 0; i < repeats; i++ {
-		s_encoded = base64.encode(s_original)
-		s+= s_encoded.len
+		resultsize := base64.encode_in_buffer(s_original, mut ebuffer)
+		s += resultsize
+		assert resultsize == s_encoded.len
 	}
 
+	dbuffer := malloc( s_decoded.len )
 	for i := 0; i < repeats; i++ {
-		s_decoded = base64.decode(s_encoded)
-		s+= s_decoded.len
+		resultsize := base64.decode_in_buffer(s_encoded, mut dbuffer)
+		s += resultsize
+		assert resultsize == s_decoded.len
 	}
-	
-	assert s_encoded.len > s_original.len
-	assert s_original == s_decoded	
+
+	println( 'Final s: $s' )
+	//	assert s == 39147008
 }