Implement base64.encode_in_buffer and base64.decode_in_buffer .
With this PR:
```shell
0[17:00:01] /v/benchmarks/base64 LINUX $
0[17:00:01] /v/benchmarks/base64 LINUX $ /v/vmaster/v -prod -cc clang-7 -o base64_v_clang_vmaster test.v
0[17:00:03] /v/benchmarks/base64 LINUX $ /v/newv/v  -prod -cc clang-7 -o base64_v_clang_newv test.v
0[17:00:21] /v/benchmarks/base64 LINUX $ /v/newv/v  -prod -cc clang-7 -o base64_v_clang_newv_2 test2.v
0[17:00:30] /v/benchmarks/base64 LINUX $
0[17:00:31] /v/benchmarks/base64 LINUX $ ../xtime.rb  ./base64_v_clang_vmaster
encode aaaa... to YWFh...:   1498 ms | 1431666688
decode YWFh... to aaaa...:   2706 ms | 1073741824
4.30s, 2418.6Mb
0[17:00:38] /v/benchmarks/base64 LINUX $ ../xtime.rb  ./base64_v_clang_newv
encode aaaa... to YWFh...:   1389 ms | 1431666688
decode YWFh... to aaaa...:   2553 ms | 1073741824
4.04s, 2430.9Mb
0[17:00:50] /v/benchmarks/base64 LINUX $ ../xtime.rb  ./base64_v_clang_newv_2
encode aaaa... to YWFh...:   1017 ms | 1431666688
decode YWFh... to aaaa...:   2307 ms | decode: 1073741824
  3325 ms | <=== total time spent running base64 encode/decode ops
 ok, fail, total =     2,     0,     2
3.33s, 2.3Mb
0[17:00:56] /v/benchmarks/base64 LINUX $
```
NB: the memory consumption of the last example is reduced, because no additional memory allocations are made, just base64 encoding/decoding .

The benchmark program test.v is from https://github.com/kostya/benchmarks .

test2.v is:
```v
import encoding.base64
import benchmark

fn main() {
  str_size := 131072
  tries := 8192

  str := 'a'.repeat(str_size)

  str2 := base64.encode(str)
  str3 := base64.decode(str2)

  str2_buffer := malloc( str2.len )
  str3_buffer := malloc( str3.len )

  print('encode ${str.substr(0, 4)}... to ${str2.substr(0, 4)}...: ')

  mut bench := benchmark.new_benchmark()
  bench.step()
  mut s := 0
  for i := 0; i < tries; i++ {
    s += base64.encode_in_buffer(str, mut str2_buffer)
  }
  bench.ok()
  println(bench.step_message('$s'))

  print('decode ${str2.substr(0, 4)}... to ${str3.substr(0, 4)}...: ')

  bench.step()
  s = 0
  for i := 0; i < tries; i++ {
    s += base64.decode_in_buffer( str2, mut str3_buffer )
  }
  bench.ok()
  println(bench.step_message('decode: $s'))

  bench.stop()
  println(bench.total_message('running base64 encode/decode ops'))
}
```