@spytheman 2019-10-26
You are right, decoding can be made faster, since Index is a static array too. Thanks!