@spytheman 2019-10-26
With the last change:
```shell
0[17:52:46] /v/benchmarks/base64 LINUX $ /v/newv/v  -keep_c -prod -cc clang-7 -o base64_v_clang_newv_2 test2.v
0[17:52:49] /v/benchmarks/base64 LINUX $ ../xtime.rb  ./base64_v_clang_newv_2
encode aaaa... to YWFh...:   1014 ms | 1431666688
decode YWFh... to aaaa...:   1963 ms | decode: 1073741824
  2977 ms | <=== total time spent running base64 encode/decode ops
 ok, fail, total =     2,     0,     2
2.98s, 2.3Mb
```
It is still slower than encoding, but I do not see easy micro optimizations left... Probably a change of the algo may be needed to make it even faster.